+++ src/user/tambahan/admin/sidang.php
@@ -17,7 +17,21 @@
                                 <li><span class="bread-blod">Jadwal</span></li>
                             </ul>
                         </div>
-                    </div>
+                    </div><br>
+                    <?php
+                    if (isset($_SESSION['success_message'])) {
+                        echo '<div class="alert alert-success" role="alert" id="successMessage">' . $_SESSION['success_message'] . '</div>';
+                        unset($_SESSION['success_message']); // Hapus pesan setelah ditampilkan
+                    }
+                    ?>
+                    <script>
+                        setTimeout(function() {
+                            var successMessage = document.getElementById('successMessage');
+                            if (successMessage) {
+                                successMessage.style.display = 'none';
+                            }
+                        }, 3000); // Menghilangkan pesan setelah 3 detik (3000 ms)
+                    </script>
                 </div>
             </div>
         </div>
@@ -26,17 +40,13 @@
 
 <div class="container-fluid">
     <div class="panel panel">
-
         <div class="panel-heading">
             <h3 class="panel-title">Jadwal Sidang</h3>
         </div>
         <div class="panel-body">
-
-
             <div class="pull-right">
                 <a href="sidang_tambah.php" class="btn btn-primary"><i class="fa fa-plus"></i> Tambah Jadwal</a>
             </div>
-
             <br>
             <br>
             <br>
@@ -57,12 +67,12 @@
                     </tr>
                 </thead>
                 <tbody>
-                    <?php 
+                    <?php
                     include '../koneksi.php';
                     $no = 1;
-                    $sidang = mysqli_query($koneksi,"SELECT * FROM sidang,sengketa WHERE sengketa_id=sengketa ORDER BY sidang_id DESC");
-                    while($p = mysqli_fetch_array($sidang)){
-                        ?>
+                    $sidang = mysqli_query($koneksi, "SELECT * FROM sidang,sengketa WHERE sengketa_id=sengketa ORDER BY sidang_id DESC");
+                    while ($p = mysqli_fetch_array($sidang)) {
+                    ?>
                         <tr>
                             <td><?php echo $no++; ?></td>
                             <td><?php echo $p['nama_pemohon'] ?></td>
@@ -72,10 +82,9 @@
                                 <b>Sidang ke-</b> : <?php echo $p['sidang_ke'] ?><br>
                             </td>
                             <td>
-                                <?php 
+                                <?php
                                 // Mengubah data JSON ke dalam bentuk array
                                 $staffArray = json_decode($p['staff_sidang'], true);
-
                                 // Mendapatkan nama staff dari tabel staff berdasarkan id_staff yang ada dalam data JSON
                                 $staffNames = array();
                                 $i = 0;
@@ -91,41 +100,33 @@
                                     $i++;
                                 }
                                 // ...
-
                                 // Menampilkan nama-nama staff yang terlibat dalam mediasi
                                 echo implode(", ", $staffNames);
-                                ?>    
+                                ?>
                             </td>
                             <td><?php echo $p['catatan_hasil'] ?></td>
                             <td><?php echo $p['keterangan_sidang'] ?></td>
                             <td><?php echo $p['agenda_sidang'] ?></td>
                             <td><?php echo $p['majelis_komisioner'] ?></td>
-
                             <td class="text-center">
-                                
                                 <!-- Tinjau button with a value -->
                                 <a target="_blank" href="sidang_preview.php?id=<?php echo $p['sidang_id']; ?>" class="btn btn-default"><i class="fa fa-search"></i> Preview</a>
                             </td>
                             <td class="text-center">
-                                    <div class="btn-group">
-                                        <a href="sidang_edit.php?id=<?php echo $p['sidang_id']; ?>" class="btn btn-default"><i class="fa fa-wrench"></i></a>
-                                        <a href="sidang_hapus.php?id=<?php echo $p['sidang_id']; ?>" class="btn btn-default"><i class="fa fa-trash"></i></a>
-                                    </div>
-                                
+                                <div class="btn-group">
+                                    <a href="sidang_edit.php?id=<?php echo $p['sidang_id']; ?>" class="btn btn-default"><i class="fa fa-wrench"></i></a>
+                                    <a href="sidang_hapus.php?id=<?php echo $p['sidang_id']; ?>"  onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?')" class="btn btn-default"><i class="fa fa-trash"></i></a>
+                                </div>
                             </td>
                         </tr>
-                        <?php 
+                    <?php
                     }
                     ?>
                 </tbody>
             </table>
-
-
         </div>
-
     </div>
 </div>
 <br><br><br><br><br><br>
 <div>
-<?php include 'footer.php'; ?></div>
-
+    <?php include 'footer.php'; ?></div>
